+++ cognite_toolkit/_cdf_tk/validation/read_yaml.py
@@ -13,7 +13,7 @@ from typing import Any, get_origin
 from cognite.client.data_classes._base import CogniteObject
 from cognite.client.utils._text import to_camel_case, to_snake_case
 
-from cognite_toolkit._cdf_tk._cls_parameters.get_type_hints import _TypeHints
+from cognite_toolkit._cdf_tk._parameters.get_type_hints import _TypeHints
 
 from .warning import (
     DataSetMissingWarning,
