+++ tests/tests_unit/test_cdf_tk/test_cls_parameters/test_functions.py
@@ -1,7 +1,7 @@
 import pytest
 from cognite.client.data_classes.data_modeling import ContainerApply, SpaceApply
 
-from cognite_toolkit._cdf_tk._cls_parameters import (
+from cognite_toolkit._cdf_tk._parameters import (
     ParameterSet,
     ParameterSpec,
     ParameterSpecSet,
